fix earthquakes in perspective view
Use higher precision for matrix calculations so that the map does not
jump around while zooming.

+++ src/mbgl/util/mat2.cpp
@@ -33,8 +33,8 @@ void matrix::identity(mat2& out) {
     out[3] = 1.0f;
 }
 
-void matrix::rotate(mat2& out, const mat2& a, float rad) {
-    float a0 = a[0], a1 = a[1], a2 = a[2], a3 = a[3],
+void matrix::rotate(mat2& out, const mat2& a, double rad) {
+    double a0 = a[0], a1 = a[1], a2 = a[2], a3 = a[3],
         s = std::sin(rad),
         c = std::cos(rad);
     out[0] = a0 *  c + a2 * s;
@@ -43,8 +43,8 @@ void matrix::rotate(mat2& out, const mat2& a, float rad) {
     out[3] = a1 * -s + a3 * c;
 };
 
-void matrix::scale(mat2& out, const mat2& a, float v0, float v1) {
-    float a0 = a[0], a1 = a[1], a2 = a[2], a3 = a[3];
+void matrix::scale(mat2& out, const mat2& a, double v0, double v1) {
+    double a0 = a[0], a1 = a[1], a2 = a[2], a3 = a[3];
     out[0] = a0 * v0;
     out[1] = a1 * v0;
     out[2] = a2 * v1;
