+++ src/mbgl/util/mat3.hpp
@@ -27,14 +27,14 @@
 
 namespace mbgl {
 
-typedef std::array<float, 9> mat3;
+typedef std::array<double, 9> mat3;
 
 namespace matrix {
 
 void identity(mat3& out);
-void translate(mat3& out, const mat3& a, float x, float y);
-void rotate(mat3& out, const mat3& a, float rad);
-void scale(mat3& out, const mat3& a, float x, float y);
+void translate(mat3& out, const mat3& a, double x, double y);
+void rotate(mat3& out, const mat3& a, double rad);
+void scale(mat3& out, const mat3& a, double x, double y);
 
 }
 }
