+++ src/mbgl/util/mat4.cpp
@@ -45,8 +45,8 @@ void matrix::identity(mat4& out) {
     out[15] = 1.0f;
 }
 
-void matrix::ortho(mat4& out, float left, float right, float bottom, float top, float near, float far) {
-    float lr = 1.0f / (left - right),
+void matrix::ortho(mat4& out, double left, double right, double bottom, double top, double near, double far) {
+    double lr = 1.0f / (left - right),
           bt = 1.0f / (bottom - top),
           nf = 1.0f / (near - far);
     out[0] = -2.0f * lr;
@@ -67,8 +67,8 @@ void matrix::ortho(mat4& out, float left, float right, float bottom, float top, 
     out[15] = 1.0f;
 }
 
-void matrix::perspective(mat4& out, float fovy, float aspect, float near, float far) {
-    float f = 1.0f / std::tan(fovy / 2.0f),
+void matrix::perspective(mat4& out, double fovy, double aspect, double near, double far) {
+    double f = 1.0f / std::tan(fovy / 2.0f),
           nf = 1.0f / (near - far);
     out[0] = f / aspect;
     out[1] = 0.0f;
@@ -107,14 +107,14 @@ void matrix::copy(mat4& out, const mat4& a) {
     out[15] = a[15];
 }
 
-void matrix::translate(mat4& out, const mat4& a, float x, float y, float z) {
+void matrix::translate(mat4& out, const mat4& a, double x, double y, double z) {
     if (&a == &out) {
         out[12] = a[0] * x + a[4] * y + a[8] * z + a[12];
         out[13] = a[1] * x + a[5] * y + a[9] * z + a[13];
         out[14] = a[2] * x + a[6] * y + a[10] * z + a[14];
         out[15] = a[3] * x + a[7] * y + a[11] * z + a[15];
     } else {
-        float a00, a01, a02, a03,
+        double a00, a01, a02, a03,
               a10, a11, a12, a13,
               a20, a21, a22, a23;
 
@@ -133,8 +133,8 @@ void matrix::translate(mat4& out, const mat4& a, float x, float y, float z) {
     }
 }
 
-void matrix::rotate_x(mat4& out, const mat4& a, float rad) {
-    float s = std::sin(rad),
+void matrix::rotate_x(mat4& out, const mat4& a, double rad) {
+    double s = std::sin(rad),
         c = std::cos(rad),
         a10 = a[4],
         a11 = a[5],
@@ -167,8 +167,8 @@ void matrix::rotate_x(mat4& out, const mat4& a, float rad) {
     out[11] = a23 * c - a13 * s;
 }
 
-void matrix::rotate_z(mat4& out, const mat4& a, float rad) {
-    float s = std::sin(rad),
+void matrix::rotate_z(mat4& out, const mat4& a, double rad) {
+    double s = std::sin(rad),
           c = std::cos(rad),
           a00 = a[0],
           a01 = a[1],
@@ -201,7 +201,7 @@ void matrix::rotate_z(mat4& out, const mat4& a, float rad) {
     out[7] = a13 * c - a03 * s;
 }
 
-void matrix::scale(mat4& out, const mat4& a, float x, float y, float z) {
+void matrix::scale(mat4& out, const mat4& a, double x, double y, double z) {
     out[0] = a[0] * x;
     out[1] = a[1] * x;
     out[2] = a[2] * x;
@@ -221,13 +221,13 @@ void matrix::scale(mat4& out, const mat4& a, float x, float y, float z) {
 }
 
 void matrix::multiply(mat4& out, const mat4& a, const mat4& b) {
-    float a00 = a[0], a01 = a[1], a02 = a[2], a03 = a[3],
+    double a00 = a[0], a01 = a[1], a02 = a[2], a03 = a[3],
           a10 = a[4], a11 = a[5], a12 = a[6], a13 = a[7],
           a20 = a[8], a21 = a[9], a22 = a[10], a23 = a[11],
           a30 = a[12], a31 = a[13], a32 = a[14], a33 = a[15];
 
     // Cache only the current line of the second matrix
-    float b0 = b[0], b1 = b[1], b2 = b[2], b3 = b[3];
+    double b0 = b[0], b1 = b[1], b2 = b[2], b3 = b[3];
     out[0] = b0 * a00 + b1 * a10 + b2 * a20 + b3 * a30;
     out[1] = b0 * a01 + b1 * a11 + b2 * a21 + b3 * a31;
     out[2] = b0 * a02 + b1 * a12 + b2 * a22 + b3 * a32;
